Update README.md
The data that we are going to use in this example is about cars. Imagine a scenario where a car is being sold, but the right price is uncertain. The aim is to maximize profit while ensuring a reasonable price for potential buyers. In this case, data science can be applied to provide a solution.